# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger Example API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/config": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Config",
                "summary": "Get Config",
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":{\"peers\":null,\"prometheus\":[\"192.168.101.105:19090\"],\"alertManagers\":null}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update Config",
                "summary": "Update Config",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateConfigReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"success\",\"data\":nil}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/deploy/ck": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deploy clickhouse",
                "summary": "Deploy clickhouse",
                "parameters": [
                    {
                        "description": "request body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.DeployCkReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"success\",\"data\":nil}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/metric/query": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Query",
                "summary": "Query",
                "parameters": [
                    {
                        "type": "string",
                        "default": "ClickHouseMetrics_Read",
                        "description": "metric name",
                        "name": "metric",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1606290000",
                        "description": "metric time",
                        "name": "time",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":[{\"metric\":{\"__name__\":\"ClickHouseMetrics_Read\",\"instance\":\"192.168.101.105:9363\",\"job\":\"clickhouse_exporter\"},\"value\":[1606290000,\"2\"]}]}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/metric/query_range": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Query Range",
                "summary": "Query Range",
                "parameters": [
                    {
                        "type": "string",
                        "default": "ClickHouseMetrics_Read",
                        "description": "metric name",
                        "name": "metric",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1606290000",
                        "description": "start time",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "1606290120",
                        "description": "end time",
                        "name": "end",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "60",
                        "description": "step window",
                        "name": "step",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"msg\":\"ok\",\"data\":[{\"metric\":{\"__name__\":\"ClickHouseMetrics_Read\",\"instance\":\"192.168.101.105:9363\",\"job\":\"clickhouse_exporter\"},\"values\":[[1606290000,\"2\"],[1606290060,\"2\"],[1606290120,\"2\"]]}]}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CkDeployConfig": {
            "type": "object",
            "properties": {
                "ckTcpPort": {
                    "type": "integer",
                    "example": 9000
                },
                "clusterName": {
                    "type": "string",
                    "example": "test"
                },
                "packageVersion": {
                    "type": "string",
                    "example": "20.8.5.45"
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "path": {
                    "type": "string",
                    "example": "/data01/"
                },
                "shards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CkShard"
                    }
                },
                "user": {
                    "type": "string",
                    "example": "ck"
                },
                "zkNodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "192.168.101.102",
                        "192.168.101.105",
                        "192.168.101.107"
                    ]
                },
                "zkPort": {
                    "type": "integer",
                    "example": 2181
                }
            }
        },
...
        "model.CkReplica": {
            "type": "object",
            "properties": {
                "ip": {
                    "type": "string",
                    "example": "192.168.101.105"
                }
            }
        },
        "model.CkShard": {
            "type": "object",
            "properties": {
                "replicas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.CkReplica"
                    }
                }
            }
        },
...
        "model.DeployCkReq": {
            "type": "object",
            "properties": {
                "clickhouse": {
                    "$ref": "#/definitions/model.CkDeployConfig"
                },
                "hosts": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "192.168.101.105"
                    ]
                },
                "password": {
                    "type": "string",
                    "example": "123456"
                },
                "user": {
                    "type": "string",
                    "example": "root"
                }
            }
        },
...
        "model.UpdateConfigReq": {
            "type": "object",
            "properties": {
                "alertManagers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "192.168.101.105:19093"
                    ]
                },
                "peers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "192.168.21.74"
                    ]
                },
                "prometheus": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "192.168.101.105:19090"
                    ]
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}